build: move the editing-command module out of ui/ to editing/commands/
ObjectCommandController, UndoBatcher and the per-aggregate edit services are Qt-free
gecko_editing code that lived under ui/. Move ui/editing/ -> editing/commands/ so the
path matches the layer. Pure relocation plus include-path updates; no behavior change.
Full build + 238 tests pass.
