+++ src/editing/commands/InventoryEditService.h
