+++ src/editing/commands/MapEditService.cpp
@@ -3,8 +3,8 @@
 #include "format/map/Map.h"
 #include "format/map/MapObject.h"
 #include "format/map/Tile.h"
-#include "ui/editing/ScriptEditService.h"
-#include "ui/editing/UndoBatcher.h"
+#include "editing/commands/ScriptEditService.h"
+#include "editing/commands/UndoBatcher.h"
 #include "util/UndoStack.h"
 
 namespace geck {
