+++ src/editing/commands/ObjectCommandController.cpp
