+++ src/editing/commands/ScriptEditService.cpp
@@ -2,7 +2,7 @@
 
 #include "format/map/Map.h"
 #include "format/map/MapObject.h"
-#include "ui/editing/UndoBatcher.h"
+#include "editing/commands/UndoBatcher.h"
 #include "util/UndoStack.h"
 
 #include <algorithm>
