+++ src/editing/commands/ScriptEditService.h
