+++ src/editing/commands/UndoBatcher.cpp
