+++ src/editing/commands/UndoBatcher.h
